Qt4 port: Replace classes Q3ListView and Q3ListViewItem by QTreeWidget and QTreeWidgetItem, respectively. Change return type of FFuListViewItem::getItemText() to std::string. Remove unnecessary default values in FFuQtListViewItem declarations. Rewrite FFuListViewItem::getDepth() as non-virtual recursive method. Add virtual method FFuListViewItem::copyTexts() for the convenience and remove unnecessary virtual method FFuListView::getNColumns().

+++ src/vpmUI/vpmUITopLevels/FuiObjectBrowser.C
@@ -202,7 +202,7 @@ void FuiObjectBrowser::onSearchViewSelectionChanged()
   std::vector<FFuListViewItem*> lvItems = searchView->getSelectedListItems();
   if (lvItems.empty()) return;
 
-  int nBaseId = atol(lvItems.front()->getItemText(0));
+  int nBaseId = std::stoi(lvItems.front()->getItemText(0));
   FmSimulationModelBase* obj = dynamic_cast<FmSimulationModelBase*>(FmDB::findObject(nBaseId));
   if (obj) outputMemo->setAllText(obj->getObjectInfo().c_str());
 }
